Moved SARIResult to core folder.
