+++ src/main/java/magpiebridge/core/SARIFResult.java
@@ -1,9 +1,7 @@
-package magpiebridge.core.analysis.configuration;
+package magpiebridge.core;
 
 import com.ibm.wala.cast.tree.CAstSourcePositionMap.Position;
 import com.ibm.wala.util.collections.Pair;
-import magpiebridge.core.AnalysisResult;
-import magpiebridge.core.Kind;
 import org.eclipse.lsp4j.DiagnosticSeverity;
 
 /** This class defines the analysis result from SARIF file */
